Fix type of /defs/lithology_attributes t_units field

+++ v2/definitions/lithology_attributes.ts
@@ -6,13 +6,13 @@ module.exports = function (req, res, next, cb) {
     return larkin.info(req, res, next);
   }
 
-  var sql = `SELECT lith_atts.id AS lith_att_id, 
-      lith_att AS name, 
-      att_type AS type, 
-      COUNT(DISTINCT unit_liths.unit_id) AS t_units 
-      FROM macrostrat.lith_atts 
-      LEFT JOIN macrostrat.unit_liths_atts ON unit_liths_atts.lith_att_id = lith_atts.id 
-      LEFT JOIN macrostrat.unit_liths ON unit_liths_atts.unit_lith_id = unit_liths.id 
+  var sql = `SELECT lith_atts.id AS lith_att_id,
+      lith_att AS name,
+      att_type AS type,
+      COUNT(DISTINCT unit_liths.unit_id)::integer AS t_units
+      FROM macrostrat.lith_atts
+      LEFT JOIN macrostrat.unit_liths_atts ON unit_liths_atts.lith_att_id = lith_atts.id
+      LEFT JOIN macrostrat.unit_liths ON unit_liths_atts.unit_lith_id = unit_liths.id
       `,
     params = {};
